pyproject.toml
[tool.poetry]
name = "or-ttask-scrap"
version = "0.1.0"
description = ""
authors = ["Oleksii <nick1nick@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.3"
gspread = "^6.1.0"
oauth2client = "^4.1.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
